Have hazards use `StrikeData` objects
