+++ src/module/actor/sheet/base.ts
@@ -1,6 +1,6 @@
 import type { ActorPF2e, CharacterPF2e } from "@actor";
 import { ActorDataPF2e } from "@actor/data";
-import { RollFunction } from "@actor/data/base";
+import { RollFunction, StrikeData } from "@actor/data/base";
 import { SAVE_TYPES } from "@actor/values";
 import { Coins, createConsumableFromSpell, DENOMINATIONS, ItemPF2e, PhysicalItemPF2e } from "@item";
 import { ItemSourcePF2e } from "@item/data";
@@ -179,6 +179,20 @@ abstract class ActorSheetPF2e<TActor extends ActorPF2e> extends ActorSheet<TActo
         );
     }
 
+    protected getStrikeFromDOM(target: HTMLElement): StrikeData | null {
+        const actionIndex = Number(target.closest<HTMLElement>("[data-action-index]")?.dataset.actionIndex);
+        const rootAction = this.actor.system.actions?.[actionIndex];
+        if (!rootAction) return null;
+
+        const altUsage = tupleHasValue(["thrown", "melee"] as const, target.dataset.altUsage)
+            ? target.dataset.altUsage
+            : null;
+
+        return altUsage
+            ? rootAction.altUsages?.find((s) => (altUsage === "thrown" ? s.item.isThrown : s.item.isMelee)) ?? null
+            : rootAction;
+    }
+
     /* -------------------------------------------- */
     /*  Event Listeners and Handlers                */
     /* -------------------------------------------- */
@@ -248,6 +262,36 @@ abstract class ActorSheetPF2e<TActor extends ActorPF2e> extends ActorSheet<TActo
             } else {
                 this.actor.rollAttribute(event, key);
             }
+        });
+
+        // Strikes
+        const $strikesList = $html.find("ol.strikes-list");
+
+        $strikesList.find("button[data-action=strike-damage]").on("click", async (event) => {
+            await this.getStrikeFromDOM(event.currentTarget)?.damage?.({ event });
+        });
+
+        $strikesList.find("button[data-action=strike-critical]").on("click", async (event) => {
+            await this.getStrikeFromDOM(event.currentTarget)?.critical?.({ event });
+        });
+
+        const attackSelectors = ".item-image[data-action=strike-attack], button[data-action=strike-attack]";
+        $strikesList.find(attackSelectors).on("click", (event) => {
+            if (!("actions" in this.actor.system)) {
+                throw ErrorPF2e("Strikes are not supported on this actor");
+            }
+
+            const target = event.currentTarget;
+            const altUsage = tupleHasValue(["thrown", "melee"] as const, target.dataset.altUsage)
+                ? target.dataset.altUsage
+                : null;
+
+            const strike = this.getStrikeFromDOM(event.currentTarget);
+            if (!strike) return;
+            const $button = $(event.currentTarget);
+            const variantIndex = Number($button.attr("data-variant-index"));
+
+            strike.variants[variantIndex]?.roll({ event, altUsage });
         });
 
         // Remove Spell Slot
