+++ static/templates/actors/npc/partials/attack.html
@@ -1,4 +1,4 @@
-<li class="item attack flexrow" data-action-index="{{actionId}}" data-item-name="{{action.item.name}}" data-item-id="{{action.sourceId}}">
+<li class="item attack flexrow" data-action-index="{{index}}" data-item-name="{{action.item.name}}" data-item-id="{{action.sourceId}}">
     <div class="attack-header">
         <h4 class="attack-name tags">
             <span>
@@ -6,7 +6,7 @@
                 <span class="activity-icon">{{action.glyph}}</span>
                 {{action.label}}
             </span>
-            {{#each traits as |trait|}}
+            {{#each action.traits as |trait|}}
                 <div class="tag tag_transparent" title="{{localize trait.description}}">{{localize trait.label}}</div>
             {{/each}}
             {{#each action.additionalEffects as |effect|}}
